web/requirements.txt
Django==2.1.7
djangorestframework==3.9.2
djangorestframework-stubs==0.3.0
psycopg2-binary==2.7.7
pytz==2018.9
